Extracted shared adapter code
