+++ src/adapter/marks.ts
