+++ src/adapter/protocol/events.ts
@@ -1,9 +1,9 @@
-import { flushTable, StringTable } from "../string-table";
+import { flushTable, StringTable } from "./string-table";
 import { Store } from "../../view/store/types";
 import { recordProfilerCommit } from "../../view/components/profiler/data/commits";
 import { ops2Tree } from "./operations";
 import { applyOperationsV1 } from "./legacy/operationsV1";
-import { Stats, stats2ops } from "../10/stats";
+import { Stats, stats2ops } from "../shared/stats";
 
 export enum MsgTypes {
 	ADD_ROOT = 1,
