+++ src/adapter/protocol/operations.test.ts
