+++ src/adapter/protocol/operations.ts
@@ -1,9 +1,9 @@
 import { ID, Tree } from "../../view/store/types";
-import { parseTable } from "../string-table";
+import { parseTable } from "./string-table";
 import { MsgTypes } from "./events";
 import { deepClone } from "../../view/components/profiler/flamegraph/modes/adjustNodesToRight";
 import { RenderReasonMap } from "../10/renderer/renderReasons";
-import { parseStats, ParsedStats } from "../10/stats";
+import { ParsedStats, parseStats } from "../shared/stats";
 
 /**
  * This is the heart of the devtools. Here we translate incoming events
