+++ src/adapter/protocol/string-table.ts
