+++ src/adapter/store.test.ts
@@ -1,5 +1,5 @@
 import { createStore } from "../view/store";
-import { applyOperationsV2 } from "./events/events";
+import { applyOperationsV2 } from "./protocol/events";
 import { expect } from "chai";
 import { fromSnapshot } from "./debug";
 import * as sinon from "sinon";
