+++ src/shells/inline/index.ts
@@ -2,7 +2,7 @@ import { createStore } from "../../view/store";
 export { createStore } from "../../view/store";
 import { render, h } from "preact";
 import { DevTools } from "../../view/components/Devtools";
-import { applyEvent } from "../../adapter/events/events";
+import { applyEvent } from "../../adapter/protocol/events";
 import { Store } from "../../view/store/types";
 import { PageHookName, DevtoolsToClient } from "../../constants";
